+++ src/TestSuiteStorage/Action/PortalNodeStorageTestContract.php
@@ -38,7 +38,7 @@ abstract class PortalNodeStorageTestContract extends TestCase
         $portalNodeCreateAction = $facade->getPortalNodeCreateAction();
         $portalNodeDeleteAction = $facade->getPortalNodeDeleteAction();
         $portalNodeCreateResults = $portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalA::class()),
+            new PortalNodeCreatePayload(PortalA::class(), PortalA::class),
         ]));
         $portalNodeCreateResult = $portalNodeCreateResults[0] ?? null;
 
@@ -87,7 +87,7 @@ abstract class PortalNodeStorageTestContract extends TestCase
         $portalNodeCreateAction = $facade->getPortalNodeCreateAction();
         $portalNodeDeleteAction = $facade->getPortalNodeDeleteAction();
         $portalNodeCreateResults = $portalNodeCreateAction->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalA::class()),
+            new PortalNodeCreatePayload(PortalA::class(), PortalA::class),
         ]));
         $portalNodeCreateResult = $portalNodeCreateResults[0] ?? null;
 
@@ -135,7 +135,7 @@ abstract class PortalNodeStorageTestContract extends TestCase
         $set = $facade->getPortalNodeStorageSetAction();
 
         $firstPortalNode = $portalNodeCreate->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalA::class()),
+            new PortalNodeCreatePayload(PortalA::class(), PortalA::class),
         ]))->first();
 
         static::assertInstanceOf(PortalNodeCreateResult::class, $firstPortalNode);